requirements.txt
absl-py==1.4.0
aiobotocore==1.2.0
aiodns==2.0.0
aiofiles==0.4.0
aiohttp==3.7.4.post0
aioitertools==0.7.1
alembic==1.10.2
asteval==0.9.21
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.1.0
bezier==2020.5.19
boto3==1.17.1
botocore==1.20.49
cachetools==5.3.0
cchardet==2.1.6
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cloudpickle==2.2.1
cmaes==0.9.1
colorlog==6.7.0
construct==2.10.61
coverage==6.1.2
cycler==0.10.0
databricks-cli==0.17.6
docker==6.0.1
entrypoints==0.4
fastapi==0.68.1
Flask==2.2.3
flatbuffers==23.3.3
fsspec==2023.1.0
ftfy==6.1.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
gunicorn==20.1.0
gym==0.21.0
h11==0.13.0
h5py==3.8.0
httptools==0.1.1
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.11.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
keras==2.11.0
kiwisolver==1.3.2
libclang==15.0.6.1
lightning-lite==1.8.6
lightning-utilities==0.6.0.post0
llvmlite==0.37.0
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.4.3
mdurl==0.1.2
mlflow==1.30.0
multidict==4.7.6
nest-asyncio==1.5.1
numba==0.51.2
numpy==1.21.6
numpy-quaternion==2020.9.5.14.42.2
nvsmi==0.4.2
oauthlib==3.1.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
optuna==3.1.0
osqp==0.6.2.post5
packaging==21.0
pandas==1.3.5
pathtools==0.1.2
Pillow==9.0.1
Pint==0.17
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.3
protobuf==4.22.1
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.21
pydantic==1.9.1
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.9.1
pytz==2022.7.1
PyYAML==5.4.1
qdldl==0.1.5.post0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.26.0
requests-oauthlib==1.3.0
rich==13.3.2
rsa==4.9
s3transfer==0.3.7
scipy==1.7.1
selenium==3.141.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.47
sqlparse==0.4.3
stable-baselines3==1.5.0
starlette==0.14.2
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
toml==0.10.1
torch==1.11.0+cu113
torchmetrics==0.11.1
torchvision==0.12.0+cu113
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.12.1
watchdog==0.10.4
watchgod==0.7
wcwidth==0.2.6
webbot==0.34
websocket-client==1.5.1
websockets==9.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.5.1
zipp==3.8.0
